# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.PingSwag:
    properties:
      health:
        example: ok
        type: string
    type: object
  request.CreateTodo:
    properties:
      email:
        example: Email TODO
        type: string
      name:
        example: Name TODO
        type: string
    type: object
  request.UpdateTodo:
    properties:
      email:
        example: Email TODO
        type: string
      name:
        example: Name TODO
        type: string
    type: object
  views.CreateTodoFailureSwag:
    properties:
      additional_info:
        additionalProperties:
          type: string
        example:
          error: Title is required
        type: object
      message:
        example: BAD Request
        type: string
      status:
        example: 400
        type: integer
    type: object
  views.CreateTodoPayload:
    properties:
      created_at:
        type: string
      email:
        example: Email TODO
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Name TODO
        type: string
    type: object
  views.CreateTodoSuccessSwag:
    properties:
      message:
        example: CREATE TODO SUCCESS
        type: string
      payload:
        $ref: '#/definitions/views.CreateTodoPayload'
      status:
        example: 201
        type: integer
    type: object
  views.DeleteTodoSuccessSwag:
    properties:
      message:
        example: DELETE TODO BY ID SUCCESS
        type: string
      status:
        example: 200
        type: integer
    type: object
  views.GetTodoPayload:
    properties:
      created_at:
        type: string
      email:
        example: Email TODO
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Name TODO
        type: string
    type: object
  views.GetTodoSuccessSwag:
    properties:
      message:
        example: GET TODO BY ID SUCCESS
        type: string
      payload:
        $ref: '#/definitions/views.GetTodoPayload'
      status:
        example: 200
        type: integer
    type: object
  views.GetTodosSuccessSwag:
    properties:
      message:
        example: GET ALL TODO SUCCESS
        type: string
      payload:
        items:
          $ref: '#/definitions/views.GetTodoPayload'
        type: array
      status:
        example: 200
        type: integer
    type: object
  views.UpdateTodoPayload:
    properties:
      created_at:
        type: string
      email:
        example: Email TODO
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Name TODO
        type: string
    type: object
  views.UpdateTodoSuccessSwag:
    properties:
      message:
        example: UPDATE TODO BY ID SUCCESS
        type: string
      payload:
        $ref: '#/definitions/views.UpdateTodoPayload'
      status:
        example: 200
        type: integer
    type: object
host: localhost:4444
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hacktiv8 Swagger
  version: "1.0"
paths:
  /ping:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.PingSwag'
      summary: ping example
      tags:
      - Health
  /todos:
    get:
      consumes:
      - application/json
      description: get all todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.GetTodosSuccessSwag'
      summary: Get All TODOS
      tags:
      - TODOS
    post:
      consumes:
      - application/json
      description: create todo
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateTodo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.CreateTodoSuccessSwag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/views.CreateTodoFailureSwag'
      summary: Create TODO
      tags:
      - TODOS
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: delete todos by ID
      parameters:
      - description: ToDo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.DeleteTodoSuccessSwag'
      summary: Delete TODO By ID
      tags:
      - TODOS
    get:
      consumes:
      - application/json
      description: get todos by ID
      parameters:
      - description: ToDo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.GetTodoSuccessSwag'
      summary: Get TODO By ID
      tags:
      - TODOS
    put:
      consumes:
      - application/json
      description: update todos by ID
      parameters:
      - description: ToDo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTodo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.UpdateTodoSuccessSwag'
      summary: Update TODO By ID
      tags:
      - TODOS
swagger: "2.0"